Automates prebuilt binary generation and commit
Sets up a GitHub Actions workflow to build and commit prebuilt binaries for different operating systems.

This eliminates the need for manual binary builds and uploads, ensuring that the latest binaries are always available.
The workflow builds binaries on Ubuntu and Windows, combines them, and commits them to the repository when new tags are pushed.

Also adds a gitignore rule to exclude prebuilds folder.